Refactor, split  solvers to corresponding files.
