+++ defines.hpp
@@ -1,0 +1,23 @@
+#pragma once
+
+#define DEBUGTRACEFILE
+#define PROB2D
+
+// Flags to change to "define" for debugging purposes
+#undef EXTRADEBUGTRACE
+#undef ANNEALING_TEST
+#undef DEBUGCHANGEPEN
+#undef DEBUGCALCPENALTIES
+#undef DEBUG_PRINTRESVALPROB
+#undef DEBUG_COUNTMISSING
+#undef DEBUG_HEURISTICS
+#undef DEBUG_IIMPHEUR
+#undef DEBUG_CLUSTERANALYSIS
+#undef DEBUG_CONNECTIONCOST
+#undef DEBUG_RESERVECOST
+#undef DEBUGCHECKCHANGE
+#undef DEBUG_CONNECTIONCOST2
+#undef CREATE_R_SCRIPTS
+#undef DEBUG_QA
+#undef DEBUG_PROB1D
+#undef DEBUG_PROB2D
